Enhancements to element processor and minor fixes. Add unsigned integer encoding. Fix error in primitives class. Add element processing for doubles. Add element processing for floats. Fix element processing for strings. Fix issue with processing of preceding length field. Add unit tests.

+++ core/src/main/java/org/beyene/protege/core/data/Primitive.java
@@ -29,7 +29,7 @@ public final class Primitive<T> {
 	public static final Primitive<Long> INTEGER = new Primitive<>(Long.class, Type.INTEGER);
 	public static final Primitive<String> STRING = new Primitive<>(String.class, Type.STRING);
 	public static final Primitive<Float> FLOAT = new Primitive<>(Float.class, Type.FLOAT);
-	public static final Primitive<Double> DOUBLE = new Primitive<>(Double.class, Type.FLOAT);
+	public static final Primitive<Double> DOUBLE = new Primitive<>(Double.class, Type.DOUBLE);
 
 	private static final Map<Type, Primitive<?>> types = new HashMap<>();
 	static {
